+++ src/lib.rs
@@ -1,5 +1,6 @@
 #![no_std]
 
+pub mod msp;
 pub mod pattern;
 pub mod postfx;
 pub mod state;